convert weekday in YamlConfigReader
